+++ src/utils/index.ts
@@ -16,6 +16,7 @@ export { ZhipuWizard } from './zhipuWizard';
 export { MiniMaxWizard } from './minimaxWizard';
 export { MoonshotWizard } from './moonshotWizard';
 export { DashscopeWizard } from './dashscopeWizard';
+export { TencentWizard } from './tencentWizard';
 export { JsonSchemaProvider } from './jsonSchemaProvider';
 export { RetryManager } from './retryManager';
 export { ModelInfoCache } from './modelInfoCache';